+++ src/onegov.user/onegov/user/tests/test_collections.py
@@ -19,6 +19,12 @@ def test_user_add(session):
 
     assert users.by_username(user.username).username == 'dmr@example.org'
 
+    groups = UserGroupCollection(session)
+    group = groups.add(name='group')
+
+    user = users.add('abc@example.org', 'p@ssw0rd', 'root', group=group)
+    assert users.by_username(user.username).group.name == 'group'
+
 
 def test_user_add_conflict(session):